login fixed
I noticed that there was an issue with the login feature where after logging in, it would stay on the same page without redirecting. However, the issue is now fixed. Now, when you log in, even if you click on 'Create' or 'Profile,' the system will automatically recognise and identify the user based on their login session