user profile
I created a user profile page that shows the user's username, email, bio, and profile picture (if they have one). I also added a feature where users can follow or unfollow the profile they're viewing. The button changes depending on whether they are already following the user or not.

For the content, I made it so users can toggle between viewing posts and recipes. The page defaults to showing posts, but by clicking on the recipe button, it switches to showing the user’s created recipes. Each post and recipe links to its own individual page for more details. I also used JavaScript to handle the toggling between posts and recipes, so the page dynamically switches between the two without needing to reload.